requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0

# UI dependencies
chainlit>=1.0.0

# LLM dependencies
openai>=1.0.0  # For OpenAI integration

# LangChain dependencies
langchain-core>=0.1.0

# Optional dependencies
# anthropic>=0.7.0  # For Anthropic integration
# redis>=5.0.0  # For Redis memory backend
# sqlalchemy>=2.0.0  # For database tools
# modelscope>=1.0.0  # For ModelScope integration

